token: split customer action handler for makecardtoken and payments
